Bulk update of recent changes in our frameworks.
Highlights:

Overall:
  - tops replace 'isa' with '[self class]', catching a bunch of old code we hadn't fixed manually
  - Updated project script phases to deal with changes to BuildStringsFromTarget
  - Removed a few stale use of SPI or since-deprecated API that gets rejected in Mac App Store submissions.
  - Added explicit NS_RETURNS_RETAINED annotations in 'copy' methods where copy isn't at the front of the method name (for clang 2.0).
  - Updated BuildStringsFromTarget to work for Mac targets, catch more errors, and obey string encoding preferences passed from Xcode.
  - Updated CopyLibraryResources to obey string encoding settings passed from Xcode.

OmniBase:
  - Expose OB_BUILTIN_ATOMICS_AVAILABLE and reenable on iOS since compiler emits valid atomic ops for ARM.
  - Treat code living in /Library the same as /System/Library so we don't warn about third-party frameworks.
  - Added (Debug-only) OBPatchTrap.[hm]. This includes a routine to find the code in objc_msgSend_stret() for a nil receiver and patches it to log/raise/etc like so:
		2010-11-16 18:56:43.714 OmniPlan[42103:a0f] -[NSScrollView setDocumentView:] (0x7fff8006b104, in AppKit) sent -bounds to nil, expecting struct return
  - Added more runtime KVO sanity checking; warn when there is a +keyPathsForValuesAffectingFoo but no -foo property.

OmniFoundation:
	- Updated the OFDataBuffer and compression API to improve performance (avoid copies, stdio, and optionally GC scans of resulting data).
  - New function OFStringContainsInvalidSequences() which checks for completely-bogus UTF-16 contents.
  - Add support for *.defaults resources. If found, register them just like defaults in bundle Info.plists/OFRegistrations. This allows us to split our shared defaults out of framework Info.plists and use the *.defaults files on both platforms.
  - Add convenience wrapper for CFPropertyList {de,}serialization, hiding the API difference between Mac OS X 10.6/iOS 4.x and iOS 3.x.
  - Split out NSData hash support and include that on the iOS side.

OmniUI/General:
  - New flatter toolbar button look, still a work in progress too.
  - Split out OUIParameters.h to expose knobs to our UI folks for tuning.
  - Added app-global system for managing the current popover to help avoid multiple-popover-avoidance code bloat (something similar really should be built-into UIKit).
  - Added OUICertificateTrustAlert and use it in the WebDAV integration.
  - Some cleanup in document system of previews, UTI helpers and document lifecycle.
  - Support for hiding the toolbar in OUIToolbarViewController.
  - Fixed keyboard avoidance for toggling between multiple keyboards changing the keyboard height.
  - Added helper functions for updating table view rows for selection changes.
  - Added OUIWithAppropriateLayerAnimations(^{...}) for passing UIKit animation down to CA in UIView setters that poke the layer's animated properties.

OmniUI/Inspectors:
	- New lighter look for inspector controls. Not shipped yet in any of our apps, so there are likely rough edges still.
	- Simplified and better sub-inspector model. Now sub-panes can override the selected object set.
	- Single-color swatch picker support for inspectors (like in OmniGraffle).
	- Clean up font inspector to be a bit more iWork-y (sans some of iWork's bugs).
	- Deprecated old inspector delegate methods, replacing with a version that passes the pane that is asking.
	- Fix bug that prevented inspector from being closed in some cases.
  - Switch OUIInspectorTextWell to use OUIEditableFrame for editing to avoid jumpy transition.
  - Add support for left-label/right-value mode in OUIInspectorTextWell.
  - Following iWork's lead, hide inspectors on undo/redo rather than trying to keep them up to date. Sometimes this is impossible anyway if the undo deletes the inspected object!

OmniUI/Text system:
  - Fixed issues with autocorrection and ideographic input due to unclear UITextInput documentation.
  - Minimal support for attachments in rich text. No built-in attachment classes yet, but the TextEditor example has a sample image cell.
  - Layout fixes for scaled text.
  - Avoid crashes due to UIKit passing nils to UITextInput methods.
  - Allow toggling on inspector support (for the selected text range).
  - Better support for right/center aligned text.

OmniUI/TextEditor sample:
  - Added toolbar item and supporting code to demo/debug text attachments.

OmniAppKit:
  - Add placeholder support for NSTextAttachment and NSTextAttachmentCell standins for iOS.
  - Fixed some issues with modal windows and OATargetSelection.
  - Some cleanup of the 64-bit-ization of the NSBezierPath extensions.
  - Fix bug with clicking in the last week of the year in OACalendarView giving us week 0 of the following year.

OmniFileStore:
  - Add delegate hooks and support for certificate warnings.

OmniQuartz:
  - Deprecate -[OQContentAnimatingLayer hasContentAnimations] with a note about why.
  - Add support for proper scaled rendering of UIImages.

OmniSoftwareUpdate:
  - Improved reporting of OSStatus-based errors via the new OFOSStatusDescription().

+++ Frameworks/OmniAppKit/OpenStepExtensions.subproj/NSBezierPath-OAInternal.h
@@ -1,4 +1,4 @@
-// Copyright 2006-2008, 2010 Omni Development, Inc.  All rights reserved.
+// Copyright 2006-2008, 2010-2011 Omni Development, Inc.  All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -31,10 +31,10 @@ typedef struct {
 
 extern void _parameterizeLine(NSPoint *coefficients, NSPoint startPoint, NSPoint endPoint);
 extern void _parameterizeCurve(NSPoint *coefficients, NSPoint startPoint, NSPoint endPoint, NSPoint controlPoint1, NSPoint controlPoint2);
-extern NSInteger intersectionsBetweenLineAndLine(const NSPoint *l1, const NSPoint *l2, struct intersectionInfo *results);
-extern NSInteger intersectionsBetweenCurveAndLine(const NSPoint *c, const NSPoint *a, struct intersectionInfo *results);
-extern NSInteger intersectionsBetweenCurveAndCurve(const NSPoint *c1coefficients, const NSPoint *c2coefficients, struct intersectionInfo *results);
-extern NSInteger intersectionsBetweenCurveAndSelf(const NSPoint *coefficients, struct intersectionInfo *results);
+extern unsigned intersectionsBetweenLineAndLine(const NSPoint *l1, const NSPoint *l2, struct intersectionInfo *results);
+extern unsigned intersectionsBetweenCurveAndLine(const NSPoint *c, const NSPoint *a, struct intersectionInfo *results);
+extern unsigned intersectionsBetweenCurveAndCurve(const NSPoint *c1coefficients, const NSPoint *c2coefficients, struct intersectionInfo *results);
+extern unsigned intersectionsBetweenCurveAndSelf(const NSPoint *coefficients, struct intersectionInfo *results);
 
 // Happy fun arbitrary constants.
 #define EPSILON 1e-10
